unjanked
I submitted the project on time and it worked, however, the code was quite ungood. I was experiencing a weird crash with CoreData that made no sense, despite implementing everything that I had done in previous projects the same, and we had a project requirement to use notification center.

the weird crash ended up getting worked around by recreating my fetchedresultscontroller with every update, then force reloading the table altogether. turns out the issue with that was due to trying to sort by a UUID type, which apparently CoreData doesn't happily do. I was able to reimplement the full functionality of the fetched results controller after modifying the users to use a normal int for id

the notification center requirement was merely to prove we knew how to use it, which is fine, but it really made no sense here. Yeah, it works, but it'd kinda be like using a hay wagon to tow a car; it'll work, but it isn't going to be pretty. i just took that out and directly called the code that already existed in the same file instead.